Bug fixes, improved visuals.

+++ LightRider/resources/line_fragment.glsl
@@ -1,6 +1,6 @@
 #version 430 core
 layout(location = 0) out vec3 color;
-layout(location = 3) out float material;
+layout(location = 3) out int material;
 
 in vec3 fragmentPosition;
 in vec3 fragmentColor;
@@ -8,5 +8,5 @@ in vec3 fragmentColor;
 void main()
 {
     color = fragmentColor;
-    material = 0.0f;
+    material = 0;
 }
